+++ src/storage/compactor/src/server.rs
@@ -27,7 +27,7 @@ use risingwave_object_store::object::parse_remote_object_store;
 use risingwave_pb::common::WorkerType;
 use risingwave_pb::compactor::compactor_service_server::CompactorServiceServer;
 use risingwave_rpc_client::MetaClient;
-use risingwave_storage::hummock::compactor::{CompactionExecutor, CompactorContext, Context};
+use risingwave_storage::hummock::compactor::{CompactionExecutor, CompactorContext};
 use risingwave_storage::hummock::hummock_meta_client::MonitoredHummockMetaClient;
 use risingwave_storage::hummock::{
     CompactorMemoryCollector, MemoryLimiter, SstableIdManager, SstableStore,
@@ -115,8 +115,8 @@ pub async fn compactor_serve(
         hummock_meta_client.clone(),
         storage_config.sstable_id_remote_fetch_number,
     ));
-    let context = Arc::new(Context {
-        options: storage_config,
+    let compactor_context = Arc::new(CompactorContext {
+        storage_config,
         hummock_meta_client: hummock_meta_client.clone(),
         sstable_store: sstable_store.clone(),
         compactor_metrics,
@@ -128,13 +128,10 @@ pub async fn compactor_serve(
         read_memory_limiter: memory_limiter,
         sstable_id_manager: sstable_id_manager.clone(),
         task_progress_manager: Default::default(),
-    });
-    let compactor_context = Arc::new(CompactorContext::with_config(
-        context,
-        CompactorRuntimeConfig {
+        compactor_runtime_config: Arc::new(tokio::sync::Mutex::new(CompactorRuntimeConfig {
             max_concurrent_task_number: opts.max_concurrent_task_number,
-        },
-    ));
+        })),
+    });
     let sub_tasks = vec![
         MetaClient::start_heartbeat_loop(
             meta_client.clone(),
