+++ src/storage/src/hummock/compactor/compaction_utils.rs
@@ -1,0 +1,236 @@
+// Copyright 2023 Singularity Data
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::collections::HashSet;
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::Arc;
+
+use itertools::Itertools;
+use minstant::Instant;
+use risingwave_common::constants::hummock::CompactionFilterFlag;
+use risingwave_hummock_sdk::filter_key_extractor::FilterKeyExtractorImpl;
+use risingwave_hummock_sdk::key::FullKey;
+use risingwave_hummock_sdk::key_range::KeyRange;
+use risingwave_hummock_sdk::prost_key_range::KeyRangeExt;
+use risingwave_hummock_sdk::table_stats::TableStatsMap;
+use risingwave_hummock_sdk::{HummockEpoch, KeyComparator};
+use risingwave_pb::hummock::{CompactTask, KeyRange as KeyRange_vec, LevelType};
+
+pub use super::context::CompactorContext;
+use crate::hummock::compactor::{
+    MultiCompactionFilter, StateCleanUpCompactionFilter, TtlCompactionFilter,
+};
+use crate::hummock::multi_builder::TableBuilderFactory;
+use crate::hummock::{
+    CachePolicy, HummockResult, MemoryLimiter, SstableBuilder, SstableBuilderOptions,
+    SstableIdManagerRef, SstableWriterFactory, SstableWriterOptions,
+};
+use crate::monitor::StoreLocalStatistic;
+
+pub struct RemoteBuilderFactory<F: SstableWriterFactory> {
+    pub sstable_id_manager: SstableIdManagerRef,
+    pub limiter: Arc<MemoryLimiter>,
+    pub options: SstableBuilderOptions,
+    pub policy: CachePolicy,
+    pub remote_rpc_cost: Arc<AtomicU64>,
+    pub filter_key_extractor: Arc<FilterKeyExtractorImpl>,
+    pub sstable_writer_factory: F,
+}
+
+#[async_trait::async_trait]
+impl<F: SstableWriterFactory> TableBuilderFactory for RemoteBuilderFactory<F> {
+    type Writer = F::Writer;
+
+    async fn open_builder(&self) -> HummockResult<SstableBuilder<Self::Writer>> {
+        // TODO: memory consumption may vary based on `SstableWriter`, `ObjectStore` and cache
+        let tracker = self
+            .limiter
+            .require_memory((self.options.capacity + self.options.block_capacity) as u64)
+            .await;
+        let timer = Instant::now();
+        let table_id = self.sstable_id_manager.get_new_sst_id().await?;
+        let cost = (timer.elapsed().as_secs_f64() * 1000000.0).round() as u64;
+        self.remote_rpc_cost.fetch_add(cost, Ordering::Relaxed);
+        let writer_options = SstableWriterOptions {
+            capacity_hint: Some(self.options.capacity + self.options.block_capacity),
+            tracker: Some(tracker),
+            policy: self.policy,
+        };
+        let writer = self
+            .sstable_writer_factory
+            .create_sst_writer(table_id, writer_options)?;
+        let builder = SstableBuilder::new(
+            table_id,
+            writer,
+            self.options.clone(),
+            self.filter_key_extractor.clone(),
+        );
+        Ok(builder)
+    }
+}
+
+#[derive(Default)]
+pub struct CompactionStatistics {
+    pub delta_drop_stat: TableStatsMap,
+
+    pub iter_total_key_counts: u64,
+    pub iter_drop_key_counts: u64,
+}
+
+impl CompactionStatistics {
+    pub fn delete_ratio(&self) -> Option<u64> {
+        if self.iter_total_key_counts == 0 {
+            return None;
+        }
+
+        Some(self.iter_drop_key_counts / self.iter_total_key_counts)
+    }
+}
+
+#[derive(Clone)]
+pub struct TaskConfig {
+    pub key_range: KeyRange,
+    pub cache_policy: CachePolicy,
+    pub gc_delete_keys: bool,
+    pub watermark: u64,
+    /// `stats_target_table_ids` decides whether a dropped key should be counted as table stats
+    /// change. For an divided SST as input, a dropped key shouldn't be counted if its table id
+    /// doesn't belong to this divided SST. See `Compactor::compact_and_build_sst`.
+    pub stats_target_table_ids: Option<HashSet<u32>>,
+}
+
+pub fn estimate_memory_use_for_compaction(task: &CompactTask) -> u64 {
+    let mut total_memory_size = 0;
+    for level in &task.input_ssts {
+        if level.level_type == LevelType::Nonoverlapping as i32 {
+            if let Some(table) = level.table_infos.first() {
+                total_memory_size += table.file_size * task.splits.len() as u64;
+            }
+        } else {
+            for table in &level.table_infos {
+                total_memory_size += table.file_size;
+            }
+        }
+    }
+    total_memory_size
+}
+
+pub fn build_multi_compaction_filter(compact_task: &CompactTask) -> MultiCompactionFilter {
+    use risingwave_common::catalog::TableOption;
+    let mut multi_filter = MultiCompactionFilter::default();
+    let compaction_filter_flag =
+        CompactionFilterFlag::from_bits(compact_task.compaction_filter_mask).unwrap_or_default();
+    if compaction_filter_flag.contains(CompactionFilterFlag::STATE_CLEAN) {
+        let state_clean_up_filter = Box::new(StateCleanUpCompactionFilter::new(
+            HashSet::from_iter(compact_task.existing_table_ids.clone()),
+        ));
+
+        multi_filter.register(state_clean_up_filter);
+    }
+
+    if compaction_filter_flag.contains(CompactionFilterFlag::TTL) {
+        let id_to_ttl = compact_task
+            .table_options
+            .iter()
+            .filter(|id_to_option| {
+                let table_option: TableOption = id_to_option.1.into();
+                table_option.retention_seconds.is_some()
+            })
+            .map(|id_to_option| (*id_to_option.0, id_to_option.1.retention_seconds))
+            .collect();
+
+        let ttl_filter = Box::new(TtlCompactionFilter::new(
+            id_to_ttl,
+            compact_task.current_epoch_time,
+        ));
+        multi_filter.register(ttl_filter);
+    }
+
+    multi_filter
+}
+
+pub async fn generate_splits(compact_task: &mut CompactTask, context: Arc<CompactorContext>) {
+    let sstable_infos = compact_task
+        .input_ssts
+        .iter()
+        .flat_map(|level| level.table_infos.iter())
+        .collect_vec();
+
+    let compaction_size = compact_task
+        .input_ssts
+        .iter()
+        .flat_map(|level| level.table_infos.iter())
+        .map(|table_info| table_info.file_size)
+        .sum::<u64>();
+
+    let sstable_size = (context.storage_config.sstable_size_mb as u64) << 20;
+    if compaction_size > sstable_size * 2 {
+        let mut indexes = vec![];
+        // preload the meta and get the smallest key to split sub_compaction
+        for sstable_info in sstable_infos {
+            indexes.extend(
+                context
+                    .sstable_store
+                    .sstable(sstable_info, &mut StoreLocalStatistic::default())
+                    .await
+                    .unwrap()
+                    .value()
+                    .meta
+                    .block_metas
+                    .iter()
+                    .map(|block| {
+                        let data_size = block.len;
+                        let full_key = FullKey {
+                            user_key: FullKey::decode(&block.smallest_key).user_key,
+                            epoch: HummockEpoch::MAX,
+                        }
+                        .encode();
+                        (data_size as u64, full_key)
+                    })
+                    .collect_vec(),
+            );
+        }
+        // sort by key, as for every data block has the same size;
+        indexes.sort_by(|a, b| KeyComparator::compare_encoded_full_key(a.1.as_ref(), b.1.as_ref()));
+        let mut splits: Vec<KeyRange_vec> = vec![];
+        splits.push(KeyRange_vec::new(vec![], vec![]));
+        let parallelism = std::cmp::min(
+            indexes.len() as u64,
+            context.storage_config.max_sub_compaction as u64,
+        );
+        let sub_compaction_data_size = std::cmp::max(compaction_size / parallelism, sstable_size);
+        let parallelism = compaction_size / sub_compaction_data_size;
+
+        if parallelism > 1 {
+            let mut last_buffer_size = 0;
+            let mut last_key: Vec<u8> = vec![];
+            let mut remaining_size = indexes.iter().map(|block| block.0).sum::<u64>();
+            for (data_size, key) in indexes {
+                if last_buffer_size >= sub_compaction_data_size
+                    && !last_key.eq(&key)
+                    && remaining_size > sstable_size
+                {
+                    splits.last_mut().unwrap().right = key.clone();
+                    splits.push(KeyRange_vec::new(key.clone(), vec![]));
+                    last_buffer_size = data_size;
+                } else {
+                    last_buffer_size += data_size;
+                }
+                remaining_size -= data_size;
+                last_key = key;
+            }
+            compact_task.splits = splits;
+        }
+    }
+}
