+++ src/storage/src/hummock/compactor/compactor_runner.rs
@@ -42,15 +42,15 @@ pub struct CompactorRunner {
 }
 
 impl CompactorRunner {
-    pub fn new(split_index: usize, context: &CompactorContext, task: CompactTask) -> Self {
-        let max_target_file_size = context.context.options.sstable_size_mb as usize * (1 << 20);
+    pub fn new(split_index: usize, context: Arc<CompactorContext>, task: CompactTask) -> Self {
+        let max_target_file_size = context.storage_config.sstable_size_mb as usize * (1 << 20);
         let total_file_size = task
             .input_ssts
             .iter()
             .flat_map(|level| level.table_infos.iter())
             .map(|table| table.file_size)
             .sum::<u64>();
-        let mut options: SstableBuilderOptions = context.context.options.as_ref().into();
+        let mut options: SstableBuilderOptions = context.storage_config.as_ref().into();
         options.capacity = std::cmp::min(task.target_file_size as usize, max_target_file_size);
         options.compression_algorithm = match task.compression_algorithm {
             0 => CompressionAlgorithm::None,
@@ -77,7 +77,7 @@ impl CompactorRunner {
             })
             .collect();
         let compactor = Compactor::new(
-            context.context.clone(),
+            context.clone(),
             options,
             key_range.clone(),
             CachePolicy::NotFill,
@@ -89,7 +89,7 @@ impl CompactorRunner {
         Self {
             compactor,
             compact_task: task,
-            sstable_store: context.context.sstable_store.clone(),
+            sstable_store: context.sstable_store.clone(),
             key_range,
             split_index,
         }
