+++ src/storage/src/hummock/compactor/mod.rs
@@ -14,6 +14,7 @@
 
 mod compaction_executor;
 mod compaction_filter;
+mod compaction_utils;
 mod compactor_runner;
 mod context;
 mod iterator;
@@ -23,111 +24,54 @@ pub(super) mod task_progress;
 use std::collections::{HashMap, HashSet};
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, Mutex};
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
 pub use compaction_executor::CompactionExecutor;
 pub use compaction_filter::{
     CompactionFilter, DummyCompactionFilter, MultiCompactionFilter, StateCleanUpCompactionFilter,
     TtlCompactionFilter,
 };
-pub use context::{CompactorContext, Context};
+pub use context::CompactorContext;
 use futures::future::try_join_all;
 use futures::{stream, StreamExt};
 pub use iterator::ConcatSstableIterator;
 use itertools::Itertools;
-use risingwave_common::constants::hummock::CompactionFilterFlag;
 use risingwave_hummock_sdk::compact::compact_task_to_string;
 use risingwave_hummock_sdk::filter_key_extractor::FilterKeyExtractorImpl;
 use risingwave_hummock_sdk::key::FullKey;
 use risingwave_hummock_sdk::key_range::KeyRange;
-use risingwave_hummock_sdk::prost_key_range::KeyRangeExt;
 use risingwave_hummock_sdk::table_stats::{add_table_stats_map, TableStats, TableStatsMap};
-use risingwave_hummock_sdk::{HummockEpoch, KeyComparator, LocalSstableInfo};
+use risingwave_hummock_sdk::LocalSstableInfo;
 use risingwave_pb::hummock::compact_task::TaskStatus;
 use risingwave_pb::hummock::subscribe_compact_tasks_response::Task;
-use risingwave_pb::hummock::{
-    CompactTask, CompactTaskProgress, KeyRange as KeyRange_vec, LevelType,
-    SubscribeCompactTasksResponse,
-};
+use risingwave_pb::hummock::{CompactTask, CompactTaskProgress, SubscribeCompactTasksResponse};
 use risingwave_rpc_client::HummockMetaClient;
 pub use shared_buffer_compact::compact;
 use tokio::sync::oneshot::{Receiver, Sender};
 use tokio::task::JoinHandle;
 
+pub use self::compaction_utils::{CompactionStatistics, RemoteBuilderFactory, TaskConfig};
 use self::task_progress::TaskProgress;
 use super::multi_builder::CapacitySplitTableBuilder;
-use super::{HummockResult, SstableBuilderOptions, SstableWriterOptions};
+use super::{HummockResult, SstableBuilderOptions};
+use crate::hummock::compactor::compaction_utils::{
+    build_multi_compaction_filter, estimate_memory_use_for_compaction, generate_splits,
+};
 use crate::hummock::compactor::compactor_runner::CompactorRunner;
 use crate::hummock::compactor::task_progress::TaskProgressGuard;
 use crate::hummock::iterator::{Forward, HummockIterator};
 use crate::hummock::multi_builder::{SplitTableOutput, TableBuilderFactory};
-use crate::hummock::utils::MemoryLimiter;
 use crate::hummock::vacuum::Vacuum;
 use crate::hummock::{
     validate_ssts, BatchSstableWriterFactory, CachePolicy, DeleteRangeAggregator, HummockError,
-    RangeTombstonesCollector, SstableBuilder, SstableIdManagerRef, SstableWriterFactory,
-    StreamingSstableWriterFactory,
+    RangeTombstonesCollector, SstableWriterFactory, StreamingSstableWriterFactory,
 };
 use crate::monitor::{CompactorMetrics, StoreLocalStatistic};
-
-pub struct RemoteBuilderFactory<F: SstableWriterFactory> {
-    sstable_id_manager: SstableIdManagerRef,
-    limiter: Arc<MemoryLimiter>,
-    options: SstableBuilderOptions,
-    policy: CachePolicy,
-    remote_rpc_cost: Arc<AtomicU64>,
-    filter_key_extractor: Arc<FilterKeyExtractorImpl>,
-    sstable_writer_factory: F,
-}
-
-#[async_trait::async_trait]
-impl<F: SstableWriterFactory> TableBuilderFactory for RemoteBuilderFactory<F> {
-    type Writer = F::Writer;
-
-    async fn open_builder(&self) -> HummockResult<SstableBuilder<Self::Writer>> {
-        // TODO: memory consumption may vary based on `SstableWriter`, `ObjectStore` and cache
-        let tracker = self
-            .limiter
-            .require_memory((self.options.capacity + self.options.block_capacity) as u64)
-            .await;
-        let timer = Instant::now();
-        let table_id = self.sstable_id_manager.get_new_sst_id().await?;
-        let cost = (timer.elapsed().as_secs_f64() * 1000000.0).round() as u64;
-        self.remote_rpc_cost.fetch_add(cost, Ordering::Relaxed);
-        let writer_options = SstableWriterOptions {
-            capacity_hint: Some(self.options.capacity + self.options.block_capacity),
-            tracker: Some(tracker),
-            policy: self.policy,
-        };
-        let writer = self
-            .sstable_writer_factory
-            .create_sst_writer(table_id, writer_options)?;
-        let builder = SstableBuilder::new(
-            table_id,
-            writer,
-            self.options.clone(),
-            self.filter_key_extractor.clone(),
-        );
-        Ok(builder)
-    }
-}
-
-#[derive(Clone)]
-pub struct TaskConfig {
-    pub key_range: KeyRange,
-    pub cache_policy: CachePolicy,
-    pub gc_delete_keys: bool,
-    pub watermark: u64,
-    /// `stats_target_table_ids` decides whether a dropped key should be counted as table stats
-    /// change. For an divided SST as input, a dropped key shouldn't be counted if its table id
-    /// doesn't belong to this divided SST. See `Compactor::compact_and_build_sst`.
-    pub stats_target_table_ids: Option<HashSet<u32>>,
-}
 
 /// Implementation of Hummock compaction.
 pub struct Compactor {
     /// The context of the compactor.
-    context: Arc<Context>,
+    context: Arc<CompactorContext>,
     task_config: TaskConfig,
     options: SstableBuilderOptions,
     get_id_time: Arc<AtomicU64>,
@@ -143,7 +87,7 @@ impl Compactor {
         mut compact_task: CompactTask,
         mut shutdown_rx: Receiver<()>,
     ) -> TaskStatus {
-        let context = compactor_context.context.clone();
+        let context = compactor_context.clone();
         // Set a watermark SST id to prevent full GC from accidentally deleting SSTs for in-progress
         // write op. The watermark is invalidated when this method exits.
         let tracker_id = match context.sstable_id_manager.add_watermark_sst_id(None).await {
@@ -236,7 +180,7 @@ impl Compactor {
             TaskProgressGuard::new(compact_task.task_id, context.task_progress_manager.clone());
         let delete_range_agg = match CompactorRunner::build_delete_range_iter(
             &compact_task,
-            &compactor_context.context.sstable_store,
+            &compactor_context.sstable_store,
             &mut multi_filter,
         )
         .await
@@ -251,11 +195,8 @@ impl Compactor {
         for (split_index, _) in compact_task.splits.iter().enumerate() {
             let filter = multi_filter.clone();
             let multi_filter_key_extractor = multi_filter_key_extractor.clone();
-            let compactor_runner = CompactorRunner::new(
-                split_index,
-                compactor_context.as_ref(),
-                compact_task.clone(),
-            );
+            let compactor_runner =
+                CompactorRunner::new(split_index, compactor_context.clone(), compact_task.clone());
             let del_agg = delete_range_agg.clone();
             let task_progress = task_progress_guard.progress.clone();
             let handle = tokio::spawn(async move {
@@ -328,7 +269,7 @@ impl Compactor {
     /// Fill in the compact task and let hummock manager know the compaction output ssts.
     async fn compact_done(
         compact_task: &mut CompactTask,
-        context: Arc<Context>,
+        context: Arc<CompactorContext>,
         output_ssts: Vec<CompactOutput>,
         task_status: TaskStatus,
     ) {
@@ -382,7 +323,7 @@ impl Compactor {
         type CompactionShutdownMap = Arc<Mutex<HashMap<u64, Sender<()>>>>;
         let (shutdown_tx, mut shutdown_rx) = tokio::sync::oneshot::channel();
         let stream_retry_interval = Duration::from_secs(60);
-        let task_progress = compactor_context.context.task_progress_manager.clone();
+        let task_progress = compactor_context.task_progress_manager.clone();
         let task_progress_update_interval = Duration::from_millis(1000);
         let join_handle = tokio::spawn(async move {
             let shutdown_map = CompactionShutdownMap::default();
@@ -419,7 +360,7 @@ impl Compactor {
                 };
                 drop(config);
 
-                let executor = compactor_context.context.compaction_executor.clone();
+                let executor = compactor_context.compaction_executor.clone();
                 // This inner loop is to consume stream or report task progress.
                 'consume_stream: loop {
                     let message = tokio::select! {
@@ -471,7 +412,7 @@ impl Compactor {
                                     Task::VacuumTask(vacuum_task) => {
                                         Vacuum::vacuum(
                                             vacuum_task,
-                                            context.context.sstable_store.clone(),
+                                            context.sstable_store.clone(),
                                             meta_client,
                                         )
                                         .await;
@@ -479,7 +420,7 @@ impl Compactor {
                                     Task::FullScanTask(full_scan_task) => {
                                         Vacuum::full_scan(
                                             full_scan_task,
-                                            context.context.sstable_store.clone(),
+                                            context.sstable_store.clone(),
                                             meta_client,
                                         )
                                         .await;
@@ -487,7 +428,7 @@ impl Compactor {
                                     Task::ValidationTask(validation_task) => {
                                         validate_ssts(
                                             validation_task,
-                                            context.context.sstable_store.clone(),
+                                            context.sstable_store.clone(),
                                         )
                                         .await;
                                     }
@@ -535,7 +476,7 @@ impl Compactor {
         compactor_metrics: Arc<CompactorMetrics>,
         mut iter: impl HummockIterator<Direction = Forward>,
         mut compaction_filter: impl CompactionFilter,
-    ) -> HummockResult<TableStatsMap>
+    ) -> HummockResult<CompactionStatistics>
     where
         F: TableBuilderFactory,
     {
@@ -566,8 +507,10 @@ impl Compactor {
         let mut table_stats_drop = TableStatsMap::default();
         let mut last_table_stats = TableStats::default();
         let mut last_table_id = None;
+        let mut compaction_statistics = CompactionStatistics::default();
         while iter.is_valid() {
             let iter_key = iter.key();
+            compaction_statistics.iter_total_key_counts += 1;
 
             let is_new_user_key =
                 last_key.is_empty() || iter_key.user_key != last_key.user_key.as_ref();
@@ -619,6 +562,8 @@ impl Compactor {
                 watermark_can_see_last_key = true;
             }
             if drop {
+                compaction_statistics.iter_drop_key_counts += 1;
+
                 let should_count = match task_config.stats_target_table_ids.as_ref() {
                     Some(target_table_ids) => {
                         target_table_ids.contains(&last_key.user_key.table_id.table_id)
@@ -646,14 +591,16 @@ impl Compactor {
         }
         iter.collect_local_statistic(&mut local_stats);
         local_stats.report_compactor(compactor_metrics.as_ref());
-        Ok(table_stats_drop)
+        compaction_statistics.delta_drop_stat = table_stats_drop;
+
+        Ok(compaction_statistics)
     }
 }
 
 impl Compactor {
     /// Create a new compactor.
     pub fn new(
-        context: Arc<Context>,
+        context: Arc<CompactorContext>,
         options: SstableBuilderOptions,
         key_range: KeyRange,
         cache_policy: CachePolicy,
@@ -701,7 +648,10 @@ impl Compactor {
         };
 
         let (split_table_outputs, table_stats_map) = if self.options.capacity as u64
-            > self.context.options.min_sst_size_for_streaming_upload
+            > self
+                .context
+                .storage_config
+                .min_sst_size_for_streaming_upload
         {
             self.compact_key_range_impl(
                 StreamingSstableWriterFactory::new(self.context.sstable_store.clone()),
@@ -800,7 +750,7 @@ impl Compactor {
             del_agg,
             self.task_config.key_range.clone(),
         );
-        let table_stats_map = Compactor::compact_and_build_sst(
+        let compaction_statstics = Compactor::compact_and_build_sst(
             &mut sst_builder,
             &self.task_config,
             self.context.compactor_metrics.clone(),
@@ -808,132 +758,9 @@ impl Compactor {
             compaction_filter,
         )
         .await?;
+
         let ssts = sst_builder.finish().await?;
-        Ok((ssts, table_stats_map))
-    }
-}
 
-pub fn estimate_memory_use_for_compaction(task: &CompactTask) -> u64 {
-    let mut total_memory_size = 0;
-    for level in &task.input_ssts {
-        if level.level_type == LevelType::Nonoverlapping as i32 {
-            if let Some(table) = level.table_infos.first() {
-                total_memory_size += table.file_size * task.splits.len() as u64;
-            }
-        } else {
-            for table in &level.table_infos {
-                total_memory_size += table.file_size;
-            }
-        }
-    }
-    total_memory_size
-}
-
-fn build_multi_compaction_filter(compact_task: &CompactTask) -> MultiCompactionFilter {
-    use risingwave_common::catalog::TableOption;
-    let mut multi_filter = MultiCompactionFilter::default();
-    let compaction_filter_flag =
-        CompactionFilterFlag::from_bits(compact_task.compaction_filter_mask).unwrap_or_default();
-    if compaction_filter_flag.contains(CompactionFilterFlag::STATE_CLEAN) {
-        let state_clean_up_filter = Box::new(StateCleanUpCompactionFilter::new(
-            HashSet::from_iter(compact_task.existing_table_ids.clone()),
-        ));
-
-        multi_filter.register(state_clean_up_filter);
-    }
-
-    if compaction_filter_flag.contains(CompactionFilterFlag::TTL) {
-        let id_to_ttl = compact_task
-            .table_options
-            .iter()
-            .filter(|id_to_option| {
-                let table_option: TableOption = id_to_option.1.into();
-                table_option.retention_seconds.is_some()
-            })
-            .map(|id_to_option| (*id_to_option.0, id_to_option.1.retention_seconds))
-            .collect();
-
-        let ttl_filter = Box::new(TtlCompactionFilter::new(
-            id_to_ttl,
-            compact_task.current_epoch_time,
-        ));
-        multi_filter.register(ttl_filter);
-    }
-
-    multi_filter
-}
-
-async fn generate_splits(compact_task: &mut CompactTask, context: Arc<Context>) {
-    let sstable_infos = compact_task
-        .input_ssts
-        .iter()
-        .flat_map(|level| level.table_infos.iter())
-        .collect_vec();
-
-    let compaction_size = compact_task
-        .input_ssts
-        .iter()
-        .flat_map(|level| level.table_infos.iter())
-        .map(|table_info| table_info.file_size)
-        .sum::<u64>();
-
-    let sstable_size = (context.options.sstable_size_mb as u64) << 20;
-    if compaction_size > sstable_size * 2 {
-        let mut indexes = vec![];
-        // preload the meta and get the smallest key to split sub_compaction
-        for sstable_info in sstable_infos {
-            indexes.extend(
-                context
-                    .sstable_store
-                    .sstable(sstable_info, &mut StoreLocalStatistic::default())
-                    .await
-                    .unwrap()
-                    .value()
-                    .meta
-                    .block_metas
-                    .iter()
-                    .map(|block| {
-                        let data_size = block.len;
-                        let full_key = FullKey {
-                            user_key: FullKey::decode(&block.smallest_key).user_key,
-                            epoch: HummockEpoch::MAX,
-                        }
-                        .encode();
-                        (data_size as u64, full_key)
-                    })
-                    .collect_vec(),
-            );
-        }
-        // sort by key, as for every data block has the same size;
-        indexes.sort_by(|a, b| KeyComparator::compare_encoded_full_key(a.1.as_ref(), b.1.as_ref()));
-        let mut splits: Vec<KeyRange_vec> = vec![];
-        splits.push(KeyRange_vec::new(vec![], vec![]));
-        let parallelism = std::cmp::min(
-            indexes.len() as u64,
-            context.options.max_sub_compaction as u64,
-        );
-        let sub_compaction_data_size = std::cmp::max(compaction_size / parallelism, sstable_size);
-        let parallelism = compaction_size / sub_compaction_data_size;
-
-        if parallelism > 1 {
-            let mut last_buffer_size = 0;
-            let mut last_key: Vec<u8> = vec![];
-            let mut remaining_size = indexes.iter().map(|block| block.0).sum::<u64>();
-            for (data_size, key) in indexes {
-                if last_buffer_size >= sub_compaction_data_size
-                    && !last_key.eq(&key)
-                    && remaining_size > sstable_size
-                {
-                    splits.last_mut().unwrap().right = key.clone();
-                    splits.push(KeyRange_vec::new(key.clone(), vec![]));
-                    last_buffer_size = data_size;
-                } else {
-                    last_buffer_size += data_size;
-                }
-                remaining_size -= data_size;
-                last_key = key;
-            }
-            compact_task.splits = splits;
-        }
+        Ok((ssts, compaction_statstics.delta_drop_stat))
     }
 }
